requirements.txt
python-dotenv==1.0.0
html2text==2020.1.16
langchain==0.0.132
